chore: move MethodsMap to venus-shared
